# src/main/resources/open-api/api-docs.yml
openapi: 3.0.1
info:
  title: Health Tracker API
  description: Health Tracker is a RESTfull service that allows users to tracker own
    health parameters.
  termsOfService: terms
  contact:
    name: Luka Zugaj
    email: luka.zugaj7@gmail.com
  license:
    name: GNU
    url: https://github.com/thombergs/code-examples/blob/master/LICENSE
  version: "1.0"
servers:
- url: http://localhost:8080
...
paths:
...
  /hypertensions:
    get:
      tags:
      - hypertension-entity-controller
      description: get-hypertension
      operationId: getCollectionResource-hypertension-get_1
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelHypertension'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelHypertension'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
      - hypertension-entity-controller
      description: create-hypertension
      operationId: postCollectionResource-hypertension-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypertensionRequestBody'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelHypertension'
  /hypertensions/{id}:
    get:
      tags:
      - hypertension-entity-controller
      description: get-hypertension
      operationId: getItemResource-hypertension-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelHypertension'
        "404":
          description: Not Found
    put:
      tags:
      - hypertension-entity-controller
      description: update-hypertension
      operationId: putItemResource-hypertension-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypertensionRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelHypertension'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelHypertension'
        "204":
          description: No Content
    delete:
      tags:
      - hypertension-entity-controller
      description: delete-hypertension
      operationId: deleteItemResource-hypertension-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - hypertension-entity-controller
      description: patch-hypertension
      operationId: patchItemResource-hypertension-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypertensionRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelHypertension'
        "204":
          description: No Content
...
  /users/{id}/hypertensions/{propertyId}:
    get:
      tags:
      - user-property-reference-controller
      description: get-hypertension-by-user-Id
      operationId: followPropertyReference-user-get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelHypertension'
        "404":
          description: Not Found
    delete:
      tags:
      - user-property-reference-controller
      description: delete-hypertension-by-user-Id
      operationId: deletePropertyReferenceId-user-delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
  /api/v1/hypertensions/{id}:
    get:
      tags:
      - Hypertension
      summary: Get a hypertension by id
      description: "Get a Hypertension object by specifying its id. The response is\
        \ Hypertension object with id, systolic and diastolic fields in the response\
        \ body."
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/HypertensionDTO'
      security:
      - bearer-authentication: []
    put:
      tags:
      - Hypertension
      summary: Update a hypertension by id
      description: Update a Hypertension object by specifying its id and sending object
        with systolic and diastolic fields in the request body.
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypertensionDTO'
        required: true
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
      security:
      - bearer-authentication: []
    delete:
      tags:
      - Hypertension
      summary: Delete a hypertension by id
      description: Delete a Hypertension object by specifying its id.
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
      security:
      - bearer-authentication: []
  /api/v1/body-mass-indices/{id}:
    get:
      tags:
      - Body Mass Index
      summary: Get a body mass index by id
      description: "Get a BodyMassIndex object by specifying its id. The response\
        \ is BodyMassIndex object with id, height, weight, age, gender and bmi value."
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/BodyMassIndexDTO'
      security:
      - google-oauth2: []
    put:
      tags:
      - Body Mass Index
      summary: Update a body mass index by id
      description: "Update a BodyMassIndex object by specifying its id and sending\
        \ object with height, weight, age and gender in the request body."
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyMassIndexDTO'
        required: true
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
      security:
      - google-oauth2: []
    delete:
      tags:
      - Body Mass Index
      summary: Delete a body mass index by id
      description: Delete a BodyMassIndex object by specifying its id.
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
      security:
      - google-oauth2: []
...
  /api/v1/hypertensions:
    get:
      tags:
      - Hypertension
      summary: Get a set of hypertensions
      description: "Get a set of Hypertension objects. The response is set of Hypertension\
        \ objects with id, systolic and diastolic fields in the response body."
      operationId: findAllForCurrentUser
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/HypertensionDTO'
      security:
      - bearer-authentication: []
    post:
      tags:
      - Hypertension
      summary: Create a hypertension
      description: Create a Hypertension object by sending object with systolic and
        diastolic fields in the request body.
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypertensionDTO'
        required: true
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "201":
          description: Created
      security:
      - bearer-authentication: []
  /api/v1/body-mass-indices:
    get:
      tags:
      - Body Mass Index
      summary: Get a set of body mass indices
      description: "Get a set of BodyMassIndex objects. The response is set of BodyMassIndex\
        \ objects with id, height, weight, age, gender and bmi value."
      operationId: findAllForCurrentUser_1
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/BodyMassIndexDTO'
      security:
      - google-oauth2: []
    post:
      tags:
      - Body Mass Index
      summary: Create a body mass index
      description: "Create a BodyMassIndex object by sending object with height, weight,\
        \ age and gender in the request body."
      operationId: save_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyMassIndexDTO'
        required: true
      responses:
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "201":
          description: Created
      security:
      - google-oauth2: []
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        title:
          type: string
        detail:
          type: string
        status:
          $ref: '#/components/schemas/StatusType'
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    EntityModelHypertension:
      required:
      - createdAt
      - modifiedAt
      type: object
      properties:
        systolic:
          type: integer
          format: int32
        diastolic:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        _links:
          $ref: '#/components/schemas/Links'
    PagedModelEntityModelHypertension:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            hypertensions:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelHypertension'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    CollectionModelHypertension:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            hypertensions:
              type: array
              items:
                $ref: '#/components/schemas/HypertensionResponse'
        _links:
          $ref: '#/components/schemas/Links'
...
    HypertensionRequestBody:
      required:
      - createdAt
      - modifiedAt
      type: object
      properties:
        id:
          type: integer
          format: int64
        systolic:
          type: integer
          format: int32
        diastolic:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
...
    HypertensionResponse:
      required:
      - createdAt
      - modifiedAt
      type: object
      properties:
        systolic:
          type: integer
          format: int32
        diastolic:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    HypertensionDTO:
      required:
      - diastolic
      - systolic
      type: object
      properties:
        id:
          type: integer
          format: int64
        systolic:
          type: integer
          format: int32
        diastolic:
          type: integer
          format: int32
    BodyMassIndexDTO:
      required:
      - age
      - gender
      - height
      - weight
      type: object
      properties:
        id:
          type: integer
          format: int64
        height:
          type: number
          format: double
        weight:
          type: number
          format: double
        age:
          type: integer
          format: int32
        gender:
          type: string
        value:
          type: number
          format: double
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
  securitySchemes:
    basicScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT